pyproject.toml
[project]
name = "retail-gen"
version = "0.1.0"
description = "Synthetic retail data generator for Omni Analytics demos"
requires-python = ">=3.11"
dependencies = [
    "faker>=22.0.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "pyarrow>=14.0.0",
    "python-dotenv>=1.0.0",
    "google-cloud-storage>=2.14.0",
    "google-cloud-bigquery>=3.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-tmp-files>=0.0.2",
]

[project.scripts]
retail-gen = "retail_gen.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/retail_gen"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
